+++ alicloud/resource_alicloud_security_group_test.go
@@ -275,7 +275,6 @@ func TestAccAlicloudSecurityGroup_inner_access(t *testing.T) {
 					resource.TestCheckNoResourceAttr("alicloud_security_group.foo", "tags"),
 				),
 			},
-
 			{
 				Config: testAccCheckSecurityGroupConfigInnerUpdate,
 				Check: resource.ComposeTestCheckFunc(
@@ -283,7 +282,70 @@ func TestAccAlicloudSecurityGroup_inner_access(t *testing.T) {
 					testAccCheckVpcExists("alicloud_vpc.vpc", &vpc),
 					resource.TestCheckResourceAttrSet("alicloud_security_group.foo", "vpc_id"),
 					resource.TestCheckResourceAttr("alicloud_security_group.foo", "inner_access", "true"),
-					resource.TestCheckResourceAttr("alicloud_security_group.foo", "name", "tf-testAccCheckSecurityGroupConfigInnerUpdate"),
+					resource.TestCheckResourceAttr("alicloud_security_group.foo", "name", "tf-testAccCheckSecurityGroupConfigInner"),
+					resource.TestCheckResourceAttr("alicloud_security_group.foo", "description", ""),
+					resource.TestCheckNoResourceAttr("alicloud_security_group.foo", "tags"),
+				),
+			},
+			{
+				Config: testAccCheckSecurityGroupConfigInner,
+				Check: resource.ComposeTestCheckFunc(
+					testAccCheckSecurityGroupExists("alicloud_security_group.foo", &group),
+					testAccCheckVpcExists("alicloud_vpc.vpc", &vpc),
+					resource.TestCheckResourceAttrSet("alicloud_security_group.foo", "vpc_id"),
+					resource.TestCheckResourceAttr("alicloud_security_group.foo", "inner_access", "false"),
+					resource.TestCheckResourceAttr("alicloud_security_group.foo", "name", "tf-testAccCheckSecurityGroupConfigInner"),
+					resource.TestCheckResourceAttr("alicloud_security_group.foo", "description", ""),
+					resource.TestCheckNoResourceAttr("alicloud_security_group.foo", "tags"),
+				),
+			},
+		},
+	})
+}
+
+func TestAccAlicloudSecurityGroup_inner_access_2(t *testing.T) {
+	var group ecs.DescribeSecurityGroupAttributeResponse
+	var vpc vpc.DescribeVpcAttributeResponse
+
+	resource.Test(t, resource.TestCase{
+		PreCheck: func() {
+			testAccPreCheck(t)
+		},
+		Providers:    testAccProviders,
+		CheckDestroy: testAccCheckSecurityGroupDestroy,
+		Steps: []resource.TestStep{
+			{
+				Config: testAccCheckSecurityGroupConfigInnerUpdate,
+				Check: resource.ComposeTestCheckFunc(
+					testAccCheckSecurityGroupExists("alicloud_security_group.foo", &group),
+					testAccCheckVpcExists("alicloud_vpc.vpc", &vpc),
+					resource.TestCheckResourceAttrSet("alicloud_security_group.foo", "vpc_id"),
+					resource.TestCheckResourceAttr("alicloud_security_group.foo", "inner_access", "true"),
+					resource.TestCheckResourceAttr("alicloud_security_group.foo", "name", "tf-testAccCheckSecurityGroupConfigInner"),
+					resource.TestCheckResourceAttr("alicloud_security_group.foo", "description", ""),
+					resource.TestCheckNoResourceAttr("alicloud_security_group.foo", "tags"),
+				),
+			},
+			{
+				Config: testAccCheckSecurityGroupConfigInner,
+				Check: resource.ComposeTestCheckFunc(
+					testAccCheckSecurityGroupExists("alicloud_security_group.foo", &group),
+					testAccCheckVpcExists("alicloud_vpc.vpc", &vpc),
+					resource.TestCheckResourceAttrSet("alicloud_security_group.foo", "vpc_id"),
+					resource.TestCheckResourceAttr("alicloud_security_group.foo", "inner_access", "false"),
+					resource.TestCheckResourceAttr("alicloud_security_group.foo", "name", "tf-testAccCheckSecurityGroupConfigInner"),
+					resource.TestCheckResourceAttr("alicloud_security_group.foo", "description", ""),
+					resource.TestCheckNoResourceAttr("alicloud_security_group.foo", "tags"),
+				),
+			},
+			{
+				Config: testAccCheckSecurityGroupConfigInnerUpdate,
+				Check: resource.ComposeTestCheckFunc(
+					testAccCheckSecurityGroupExists("alicloud_security_group.foo", &group),
+					testAccCheckVpcExists("alicloud_vpc.vpc", &vpc),
+					resource.TestCheckResourceAttrSet("alicloud_security_group.foo", "vpc_id"),
+					resource.TestCheckResourceAttr("alicloud_security_group.foo", "inner_access", "true"),
+					resource.TestCheckResourceAttr("alicloud_security_group.foo", "name", "tf-testAccCheckSecurityGroupConfigInner"),
 					resource.TestCheckResourceAttr("alicloud_security_group.foo", "description", ""),
 					resource.TestCheckNoResourceAttr("alicloud_security_group.foo", "tags"),
 				),
@@ -310,11 +372,10 @@ func testAccCheckSecurityGroupExists(n string, sg *ecs.DescribeSecurityGroupAttr
 		log.Printf("[WARN] security group id %#v", rs.Primary.ID)
 
 		if err != nil {
-			return err
+			return WrapError(err)
 		}
-		if d.SecurityGroupId == rs.Primary.ID {
-			*sg = d
-		}
+		*sg = d
+
 		return nil
 	}
 }
@@ -328,17 +389,15 @@ func testAccCheckSecurityGroupDestroy(s *terraform.State) error {
 			continue
 		}
 
-		group, err := ecsService.DescribeSecurityGroupAttribute(rs.Primary.ID)
+		_, err := ecsService.DescribeSecurityGroupAttribute(rs.Primary.ID)
 
 		if err != nil {
-			if NotFoundError(err) || IsExceptedErrors(err, []string{InvalidSecurityGroupIdNotFound}) {
+			if NotFoundError(err) {
 				continue
 			}
 			return err
 		}
-		if group.SecurityGroupId != "" {
-			return fmt.Errorf("Error SecurityGroup still exist")
-		}
+		return WrapError(Error("Error SecurityGroup still exist"))
 	}
 	return nil
 }
@@ -455,7 +514,7 @@ resource "alicloud_vpc" "vpc" {
 `
 const testAccCheckSecurityGroupConfigInnerUpdate = `
 variable "name" {
-  default = "tf-testAccCheckSecurityGroupConfigInnerUpdate"
+  default = "tf-testAccCheckSecurityGroupConfigInner"
 }
 resource "alicloud_security_group" "foo" {
   name = "${var.name}"